# docs/swagger.yaml
definitions:
  flow-control_internal_types.FlowEvent:
    properties:
      data:
        additionalProperties: true
        description: Data contains additional event-specific data
        type: object
      flow_id:
        description: FlowID identifies the flow that generated the event
        type: string
      message:
        description: Message provides details about the event
        type: string
      node_id:
        description: NodeID identifies the node that generated the event
        type: string
      timestamp:
        description: Timestamp indicates when the event occurred
        type: string
      type:
        description: Type indicates the kind of event
        type: string
    type: object
  flow-control_internal_types.RuntimeFlow:
    properties:
      config:
        description: Config contains the flow's configuration in JSON format
        type: string
      created_at:
        description: CreatedAt is the timestamp when the flow was created
        type: string
      description:
        description: Description provides additional details about the flow
        type: string
      id:
        description: ID uniquely identifies the flow
        type: string
      name:
        description: Name is a human-readable name for the flow
        type: string
      status:
        description: Status represents the current state of the flow
        type: string
      updated_at:
        description: UpdatedAt is the timestamp when the flow was last updated
        type: string
      version:
        description: Version is the flow's version number
        type: string
    type: object
info:
  contact: {}
paths:
  /flows:
    get:
      consumes:
      - application/json
      description: Get a list of all flows
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
            type: array
      summary: List all flows
      tags:
      - flows
    post:
      consumes:
      - application/json
      description: Create a new flow with the provided configuration
      parameters:
      - description: Flow configuration
        in: body
        name: flow
        required: true
        schema:
          $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
      summary: Create a new flow
      tags:
      - flows
  /flows/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a flow by its ID
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a flow
      tags:
      - flows
    get:
      consumes:
      - application/json
      description: Get a flow's details by its ID
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
      summary: Get a flow by ID
      tags:
      - flows
    put:
      consumes:
      - application/json
      description: Update an existing flow's configuration
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated flow configuration
        in: body
        name: flow
        required: true
        schema:
          $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flow-control_internal_types.RuntimeFlow'
      summary: Update a flow
      tags:
      - flows
  /flows/{id}/events:
    get:
      consumes:
      - application/json
      description: Get real-time events for a flow via SSE
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/flow-control_internal_types.FlowEvent'
      summary: Get flow events
      tags:
      - flows
  /flows/{id}/status:
    put:
      consumes:
      - application/json
      description: Update a flow's status
      parameters:
      - description: Flow ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update flow status
      tags:
      - flows
swagger: "2.0"