# internal/store/models.go
package store

import "flow-control/internal/types"

// RuntimeFlow represents a flow configuration and its runtime state.
// It is re-exported from the types package for convenience.
type RuntimeFlow = types.RuntimeFlow

// FlowEvent represents a real-time event from a flow.
// It is re-exported from the types package for convenience.
type FlowEvent = types.FlowEvent

// FlowMetrics represents metrics collected during flow execution.
// It is re-exported from the types package for convenience.
type FlowMetrics = types.FlowMetrics